+++ apps/fluent-tester/src/FluentTester/TestComponents/Badge/index.ts
@@ -1,0 +1,2 @@
+export * from './BadgeTest';
+export * from './consts';