+++ packages/experimental/Button/src/useButton.ts
@@ -1,18 +1,17 @@
 import * as React from 'react';
-import { useAsPressable, useKeyUpProps, useOnPressWithFocus, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
-import { ButtonPropsWithInnerRef, ButtonState } from './Button.types';
+import { useAsPressable, useKeyProps, useOnPressWithFocus, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
+import { ButtonProps, ButtonState } from './Button.types';
 import { memoize } from '@fluentui-react-native/framework';
 
-export const useButton = (props: ButtonPropsWithInnerRef): ButtonState => {
+export const useButton = (props: ButtonProps): ButtonState => {
   // attach the pressable state handlers
-  const defaultRef = React.useRef(null);
-  const { onClick, innerRef, disabled, loading, ...rest } = props;
-  const ref = innerRef !== null ? innerRef : defaultRef;
+  const defaultComponentRef = React.useRef(null);
+  const { onClick, componentRef = defaultComponentRef, disabled, loading, ...rest } = props;
   // GH #1336: Set focusRef to null if button is disabled to prevent getting keyboard focus.
-  const focusRef = disabled ? null : ref;
+  const focusRef = disabled ? null : componentRef;
   const onClickWithFocus = useOnPressWithFocus(focusRef, onClick);
   const pressable = useAsPressable({ ...rest, onPress: onClickWithFocus });
-  const onKeyUpProps = useKeyUpProps(onClick, ' ', 'Enter');
+  const onKeyUpProps = useKeyProps(onClick, ' ', 'Enter');
   const isDisabled = !!disabled || !!loading;
 
   return {
@@ -25,7 +24,7 @@ export const useButton = (props: ButtonPropsWithInnerRef): ButtonState => {
       accessibilityState: getAccessibilityState(isDisabled),
       enableFocusRing: true,
       focusable: !isDisabled,
-      ref: useViewCommandFocus(ref),
+      ref: useViewCommandFocus(componentRef),
       ...onKeyUpProps,
       iconPosition: props.iconPosition || 'before',
       loading,