+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
+This log was last generated on Wed, 26 Jan 2022 17:02:27 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.0
+
+Wed, 26 Jan 2022 17:02:27 GMT
+
+### Minor changes
+
+- Adding code that was added to usKeyProps.win32. I believe this was accidently left out of a previous PR and it broke our UWP app (safreibe@microsoft.com)
+
+## 0.11.15
+
+Tue, 25 Jan 2022 21:55:30 GMT
+
+### Patches
+
+- Default to onKeyDown on macOS (sanajmi@microsoft.com)
+
+## 0.11.14
+
+Tue, 25 Jan 2022 16:48:23 GMT
+
+### Patches
+
+- Add Keyboarding integration for Windows (safreibe@microsoft.com)
 
 ## 0.11.13